# Use wire-type compatible source in TinyGo