--- a/example/main.go
+++ b/example/main.go
@@ -10,13 +10,20 @@ func main() {
 }
 
 // nameEqualsPodSpec schedules this node if its name equals its pod spec.
-func nameEqualsPodSpec(nodeInfo api.NodeInfo, pod api.Pod) (api.StatusCode, string) {
-	nodeName := nodeInfo.Node().Metadata.Name
-	podSpecNodeName := pod.Spec().NodeName
+func nameEqualsPodSpec(args api.FilterArgs) (api.StatusCode, string) {
+	nodeName := nilToEmpty(args.NodeInfo().Node().Metadata.Name)
+	podSpecNodeName := nilToEmpty(args.Pod().Spec.NodeName)
 
 	if len(podSpecNodeName) == 0 || podSpecNodeName == nodeName {
 		return api.StatusCodeSuccess, ""
 	} else {
-		return api.StatusCodeUnschedulable, ""
+		return api.StatusCodeUnschedulable, podSpecNodeName + " != " + nodeName
 	}
+}
+
+func nilToEmpty(ptr *string) (s string) {
+	if ptr != nil {
+		s = *ptr
+	}
+	return
 }