--- a/guest/api/types.go
+++ b/guest/api/types.go
@@ -4,21 +4,25 @@ import protoapi "sigs.k8s.io/kube-scheduler-wasm-extension/kubernetes/proto/api"
 
 // Filter is a WebAssembly implementation of framework.FilterPlugin.
 type Filter interface {
-	Filter(NodeInfo, Pod) (statusCode StatusCode, statusReason string)
+	Filter(FilterArgs) (statusCode StatusCode, statusReason string)
 }
 
 // FilterFunc adapts an ordinary function to a Filter.
-type FilterFunc func(NodeInfo, Pod) (statusCode StatusCode, statusReason string)
+type FilterFunc func(FilterArgs) (statusCode StatusCode, statusReason string)
 
-// Filter returns f(n, p).
-func (f FilterFunc) Filter(n NodeInfo, p Pod) (statusCode StatusCode, statusReason string) {
-	return f(n, p)
+// Filter returns f(a).
+func (f FilterFunc) Filter(a FilterArgs) (statusCode StatusCode, statusReason string) {
+	return f(a)
+}
+
+// FilterArgs are the arguments to a Filter.
+//
+// Note: The arguments are lazy fetched to avoid overhead for properties not in use.
+type FilterArgs interface {
+	NodeInfo() NodeInfo
+	Pod() *protoapi.Pod
 }
 
 type NodeInfo interface {
-	Node() *protoapi.IoK8SApiCoreV1Node
-}
-
-type Pod interface {
-	Spec() *protoapi.IoK8SApiCoreV1PodSpec
+	Node() *protoapi.Node
 }