--- a/guest/go.mod
+++ b/guest/go.mod
@@ -1,5 +1,6 @@
 module sigs.k8s.io/kube-scheduler-wasm-extension/guest
 
+// Match highest TinyGo's supported version of Go: 1.19 as of TinyGo 0.27
 go 1.19
 
 replace sigs.k8s.io/kube-scheduler-wasm-extension/kubernetes/proto => ../kubernetes/proto
@@ -8,6 +9,5 @@ require sigs.k8s.io/kube-scheduler-wasm-extension/kubernetes/proto v0.0.0-000101
 
 require (
 	github.com/google/go-cmp v0.5.9 // indirect
-	github.com/knqyf263/go-plugin v0.7.0 // indirect
 	google.golang.org/protobuf v1.30.0 // indirect
 )