--- a/guest/internal/imports/imports_stub.go
+++ b/guest/internal/imports/imports_stub.go
@@ -10,7 +10,7 @@ func _nodeInfoNode(ptr uint32, limit bufLimit) (len uint32) {
 	return 0
 }
 
-// podSpec is stubbed for compilation outside TinyGo.
-func _podSpec(ptr uint32, limit bufLimit) (len uint32) {
+// pod is stubbed for compilation outside TinyGo.
+func _pod(ptr uint32, limit bufLimit) (len uint32) {
 	return 0
 }