--- a/kubernetes/proto/go.mod
+++ b/kubernetes/proto/go.mod
@@ -1,8 +1,6 @@
 module sigs.k8s.io/kube-scheduler-wasm-extension/kubernetes/proto
 
-go 1.20
+// Match highest TinyGo's supported version of Go: 1.19 as of TinyGo 0.27
+go 1.19
 
-require (
-	github.com/knqyf263/go-plugin v0.7.0
-	google.golang.org/protobuf v1.30.0
-)
+require google.golang.org/protobuf v1.30.0