--- a/kubernetes/proto/tools/go.mod
+++ b/kubernetes/proto/tools/go.mod
@@ -4,13 +4,7 @@ go 1.20
 
 require (
 	github.com/knqyf263/go-plugin v0.7.0
-	github.com/sanposhiho/openapi2proto v0.0.0-20230521044535-d1080a134e37
 	google.golang.org/protobuf v1.30.0
 )
 
-require (
-	github.com/dolmen-go/jsonptr v0.0.0-20190227181151-a830c2c3c0fe // indirect
-	github.com/pkg/errors v0.8.1 // indirect
-	github.com/planetscale/vtprotobuf v0.4.0 // indirect
-	gopkg.in/yaml.v2 v2.2.2 // indirect
-)
+require github.com/planetscale/vtprotobuf v0.4.0 // indirect