--- a/kubernetes/proto/tools/tools.go
+++ b/kubernetes/proto/tools/tools.go
@@ -6,6 +6,5 @@ package tools
 
 import (
 	_ "github.com/knqyf263/go-plugin/cmd/protoc-gen-go-plugin"
-	_ "github.com/sanposhiho/openapi2proto/cmd/openapi2proto"
 	_ "google.golang.org/protobuf/cmd/protoc-gen-go"
 )