… example/go.sum
… example/main.wasm
… guest/go.sum
… kubernetes/proto/api/generated.pb.go
… kubernetes/proto/api/generated_vtproto.pb.go
… kubernetes/proto/go.sum
… kubernetes/proto/instr/generated.pb.go
… kubernetes/proto/instr/generated_vtproto.pb.go
… kubernetes/proto/kubernetes.proto
… kubernetes/proto/meta/generated.pb.go
… kubernetes/proto/meta/generated_vtproto.pb.go
… kubernetes/proto/resource/generated.pb.go
… kubernetes/proto/resource/generated_vtproto.pb.go
… kubernetes/proto/runtime/generated.pb.go
… kubernetes/proto/runtime/generated_vtproto.pb.go
… kubernetes/proto/schema/generated.pb.go
… kubernetes/proto/schema/generated_vtproto.pb.go
… kubernetes/proto/tools/go.sum