#### What type of PR is this?

/kind feature

#### What this PR does / why we need it:

Before, we had incompatible types, which among other things, made the POC difficult to reason with from a performance point of view. We were manually converting fields, which would never work in practice as the model is quite large.

This uses the real protos k8s does, just generating TinyGo compatible guest code from them. This now serializes the `v1.Pod` directly from the input type in a way TinyGo can get it. To eliminate a performance hit from filters that don't access the current pod, this extracts a `FilterArgs` type. Doing so gives us flexibility in implementation, especially as we add more parameters later.

#### Which issue(s) this PR fixes:

Fixes #13

#### Special notes for your reviewer:

This change slows down execution, but that's good because it is more realistic. Before we manually copied a single field, so the *after* is a better baseline. Note: this still doesn't use realistic data, but at least we can generate it now!
```bash
$ cd plugin
$ go test -run='^$' -bench '^BenchmarkFilter$' . -count=6 > old.txt
$ go test -run='^$' -bench '^BenchmarkFilter$' . -count=6 > new.txt
$ benchstat old.txt new.txt
goos: darwin
goarch: arm64
pkg: sigs.k8s.io/kube-scheduler-wasm-extension/plugin
                                                    │   old.txt   │              new.txt               │
                                                    │   sec/op    │   sec/op     vs base               │
Filter/success:_node_is_match_with_spec.NodeName-12   12.62µ ± 0%   19.14µ ± 1%  +51.73% (p=0.002 n=6)
```