@codefromthecrypt 2023-05-28
Thanks for the fast feedback. now we are unlocked to do more modeling and performance work!